[Refactor] WriteReadAging
	- add RandomGenerator class
	- method extract to getrandomnumber
