+++ TestShell/TestShell.vcxproj.filters
@@ -82,6 +82,9 @@
     <ClCompile Include="flush_command.cpp">
       <Filter>소스 파일\cmds</Filter>
     </ClCompile>
+    <ClCompile Include="test_util.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="command_parser.h">
@@ -145,6 +148,11 @@
     <ClInclude Include="write_read_aging_command.h" />
     <ClInclude Include="flush_command.h">
       <Filter>헤더 파일\cmds</Filter>
+    </ClInclude>
+    <ClInclude Include="ssd_impl.h" />
+    <ClInclude Include="write_read_aging_command.h" />
+    <ClInclude Include="test_util.h">
+      <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
