+++ TestShell/write_read_aging_command.h
@@ -8,7 +8,6 @@ public:
 	void run(const CommandRunner& cmdRunner) const override;
 	void printHelp() const override;
 private:
-	string hexToString(unsigned int random_value) const;
 	void printProcess() const;
 };
